# docs/swagger.yaml
definitions:
...
  dto.CreateEnterpriseDTO:
    properties:
      cnpj:
        type: string
      name:
        type: string
    required:
    - cnpj
    - name
    type: object
  dto.CreateEnterpriseResponse:
    properties:
      cnpj:
        type: string
      createdAt:
        type: string
      name:
        type: string
    type: object
  dto.CreateTokenAPIDto:
    properties:
      email:
        example: admin@test.com
        type: string
      expires:
        example: "2025-12-31T23:59:59Z"
        type: string
    required:
    - email
    - expires
    type: object
  dto.CreateTokenAPIResponse:
    properties:
      token:
        example: 645fe33b232a1b0c19f4e2bf1e475df2aa381cf45457ab6885cfb9c4bcd9e288
        type: string
    type: object
...
  dto.LoginUserDTO:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginUserResponse:
    properties:
      permissions:
        $ref: '#/definitions/dto.ReadPermissionsResponse'
      token:
        type: string
      user:
        $ref: '#/definitions/dto.ReadUserResponse'
    type: object
  dto.LogoutUserDTO:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  dto.ReadEnterpriseResponse:
    properties:
      cnpj:
        type: string
      createdAt:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  dto.ReadEnterprisesResponse:
    properties:
      enterprises:
        items:
          $ref: '#/definitions/dto.ReadEnterpriseResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  dto.ReadPermissionResponse:
    properties:
      code:
        type: string
      description:
        type: string
    type: object
  dto.ReadPermissionsResponse:
    properties:
      permissions:
        items:
          $ref: '#/definitions/dto.ReadPermissionResponse'
        type: array
    type: object
  dto.ReadUserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      enterprise:
        $ref: '#/definitions/dto.ReadEnterpriseResponse'
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      number:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.RefreshTokenUserDTO:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  dto.RefreshTokenUserResponse:
    properties:
      message:
        example: Token refreshed successfully
        type: string
    type: object
  dto.UpdateEnterpriseDTO:
    properties:
      cnpj:
        type: string
      newCnpj:
        type: string
      newName:
        type: string
    required:
    - cnpj
    type: object
  dto.UpdateEnterpriseResponse:
    properties:
      newCnpj:
        type: string
      newName:
        type: string
      oldCnpj:
        type: string
      updatedAt:
        type: string
    type: object
...
  rest_err.Causes:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  rest_err.RestErr:
    properties:
      causes:
        items:
          $ref: '#/definitions/rest_err.Causes'
        type: array
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
  description: Generic administrative API for a Basic Crud
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Basic CRUD API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return user data, token and permissions
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      summary: Login
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logs user out of the system by invalidating their token
      parameters:
      - description: Logout request (only email)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LogoutUserDTO'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Reset token expiration if it's still valid
      parameters:
      - description: Refresh credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RefreshTokenUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Refresh token
      tags:
      - Auth
  /auth/token:
    post:
      consumes:
      - application/json
      description: Generates a new API token using a valid Bearer token and email
      parameters:
      - description: Token creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTokenAPIDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CreateTokenAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Generate API Token
      tags:
      - Auth
  /enterprise/v1/:
    post:
      consumes:
      - application/json
      description: Create enterprise by CNPJ and name
      parameters:
      - description: Company data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateEnterpriseDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CreateEnterpriseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Create enterprise
      tags:
      - Enterprise
    put:
      consumes:
      - application/json
      description: Update enterprise by CNPJ
      parameters:
      - description: Company data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateEnterpriseDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateEnterpriseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Update enterprise
      tags:
      - Enterprise
  /enterprise/v1/{cnpj}:
    delete:
      consumes:
      - application/json
      description: Delete an enterprise by CNPJ
      parameters:
      - description: CNPJ of the enterprise
        in: path
        name: cnpj
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Delete enterprise
      tags:
      - Enterprise
  /enterprise/v1/read:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of enterprises
      parameters:
      - description: Page number (min 1)
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReadEnterprisesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: List enterprises
      tags:
      - Enterprise
  /enterprise/v1/read/{cnpj}:
    get:
      consumes:
      - application/json
      description: Read enterprise by CNPJ
      parameters:
      - description: CNPJ of the enterprise
        in: path
        name: cnpj
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReadEnterpriseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Read enterprise
      tags:
      - Enterprise
...
securityDefinitions:
  BearerAuth:
    description: 'Enter the token in the format: Bearer <your_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"